+++ app/render/passes/ParticleUpdatePass.cpp
@@ -32,7 +32,7 @@ void prefillParticleBuffer(RenderContext* rc, AllocatedBuffer& buffer)
     gpuParticle->initialVelocity = glm::vec4(cpuParticle._initialVelocity, cpuParticle._scale);
     gpuParticle->currentPosition = glm::vec4(0.0f, 0.0f, 0.0f, cpuParticle._spawnDelay);
     gpuParticle->currentVelocity = glm::vec4(0.0f);
-    gpuParticle->renderableId = cpuParticle._renderableId;
+    gpuParticle->renderableId = static_cast<uint32_t>(cpuParticle._renderableId);
     gpuParticle->alive = 0;
   }
 
@@ -129,7 +129,7 @@ void ParticleUpdatePass::registerToGraph(FrameGraphBuilder& fgb, RenderContext* 
         1, 1, &(*exeParams.descriptorSets)[0],
         0, nullptr);
 
-      uint32_t numX = particleSize / 8;
+      uint32_t numX = static_cast<uint32_t>(particleSize) / 8;
 
       vkCmdDispatch(*exeParams.cmdBuffer, numX + 1, 1, 1);
     });
