+++ app/render/passes/SurfelUpdateRayTracingPass.cpp
@@ -18,8 +18,8 @@ void SurfelUpdateRayTracingPass::registerToGraph(FrameGraphBuilder& fgb, RenderC
   RenderPassRegisterInfo info{};
   info._name = "SurfelUpdateRT";
 
-  const int numSurfelsX = std::ceil(rc->swapChainExtent().width / 32.0);
-  const int numSurfelsY = std::ceil(rc->swapChainExtent().height / 32.0);
+  const int numSurfelsX = static_cast<int>(std::ceil(rc->swapChainExtent().width / 32.0));
+  const int numSurfelsY = static_cast<int>(std::ceil(rc->swapChainExtent().height / 32.0));
   const int octSize = 8;
 
   {
